inputdrivers: tslib: add support for ts_read_mt()
First of all, ts_read_mt() is available in newer version of tslib. With
those, also their ABI changed and we need to check for it in configure.
We drop checking for tslib-0.0 which is ancient and historic at best in
favor of the up-to-date tslib.

If supported by the libts library, ts_read_mt() is now used instead of
ts_read(). This would make it trivially easy to add real multitouch support
of it can be used inside of DirectFB.

Signed-off-by: Martin Kepplinger <martink@posteo.de>